// src/lib/api/generated/model/form.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Django Continuous Integration API
 * API documentation for the Django Continuous Integration project with tenant-aware access rules.
 * OpenAPI spec version: 1.0.0
 */
import type { Status37cEnum } from './status37cEnum';

export interface Form {
  readonly id: string;
  /** @nullable */
  readonly tenant: string | null;
  /** @nullable */
  framework?: string | null;
  readonly framework_code: string;
  /** @maxLength 255 */
  title: string;
  description?: string;
  status?: Status37cEnum;
  /** Global seed template, not tied to any tenant. */
  readonly is_template: boolean;
  readonly question_count: string;
  readonly created_by: string;
  readonly created_at: string;
  readonly updated_at: string;
}